+++ lib/nyxx.dart
@@ -18,6 +18,7 @@ export "src/core/application/oauth2_application.dart" show IOAuth2Application;
 export 'src/core/audit_logs/audit_log.dart' show IAuditLog;
 export 'src/core/audit_logs/audit_log_change.dart' show ChangeKeyType, IAuditLogChange;
 export 'src/core/audit_logs/audit_log_entry.dart' show IAuditLogEntry, AuditLogEntryType;
+export 'src/core/audit_logs/audit_log_options.dart' show IAuditLogOptions;
 export 'src/core/channel/cacheable_text_channel.dart' show ICacheableTextChannel;
 export 'src/core/channel/channel.dart' show IChannel, ChannelType;
 export 'src/core/channel/dm_channel.dart' show IDMChannel;
@@ -139,7 +140,7 @@ export 'src/internal/exceptions/invalid_shard_exception.dart' show InvalidShardE
 export 'src/internal/exceptions/invalid_snowflake_exception.dart' show InvalidSnowflakeException;
 export 'src/internal/exceptions/missing_token_error.dart' show MissingTokenError;
 export 'src/internal/exceptions/unrecoverable_nyxx_error.dart' show UnrecoverableNyxxError;
-export 'src/internal/http/http_response.dart' show IHttpResponse, IHttpResponseError, IHttpResponseSucess;
+export 'src/internal/http/http_response.dart' show IHttpResponse, IHttpResponseError, IHttpResponseSuccess;
 export 'src/internal/interfaces/convertable.dart' show Convertable;
 export 'src/internal/interfaces/disposable.dart' show Disposable;
 export 'src/internal/interfaces/message_author.dart' show IMessageAuthor;