+++ lib/src/core/audit_logs/audit_log_entry.dart
@@ -6,6 +6,7 @@ import 'package:nyxx/src/core/user/user.dart';
 import 'package:nyxx/src/internal/cache/cacheable.dart';
 import 'package:nyxx/src/typedefs.dart';
 import 'package:nyxx/src/utils/enum.dart';
+import 'package:nyxx/src/core/audit_logs/audit_log_options.dart';
 
 abstract class IAuditLogEntry implements SnowflakeEntity {
   /// Id of the affected entity (webhook, user, role, etc.)
@@ -21,7 +22,7 @@ abstract class IAuditLogEntry implements SnowflakeEntity {
   AuditLogEntryType get type;
 
   /// Additional info for certain action types
-  String? get options;
+  IAuditLogOptions? get options;
 
   /// The reason for the change
   String? get reason;
@@ -49,7 +50,7 @@ class AuditLogEntry extends SnowflakeEntity implements IAuditLogEntry {
 
   /// Additional info for certain action types
   @override
-  late final String? options;
+  late final IAuditLogOptions? options;
 
   /// The reason for the change
   @override
@@ -57,7 +58,7 @@ class AuditLogEntry extends SnowflakeEntity implements IAuditLogEntry {
 
   /// Creates an instance of [AuditLogEntry]
   AuditLogEntry(RawApiMap raw, INyxx client) : super(Snowflake(raw["id"] as String)) {
-    targetId = raw["targetId"] as String;
+    targetId = raw["target_id"] as String;
 
     changes = [
       if (raw["changes"] != null)
@@ -67,7 +68,11 @@ class AuditLogEntry extends SnowflakeEntity implements IAuditLogEntry {
     user = UserCacheable(client, Snowflake(raw["user_id"]));
     type = AuditLogEntryType._create(raw["action_type"] as int);
 
-    options = raw["options"] as String?;
+    if (raw["options"] != null) {
+      options = AuditLogOptions(raw["options"] as RawApiMap);
+    } else {
+      options = null;
+    }
 
     reason = raw["reason"] as String?;
   }
@@ -87,6 +92,9 @@ class AuditLogEntryType extends IEnum<int> {
   static const AuditLogEntryType memberBanRemove = AuditLogEntryType._create(23);
   static const AuditLogEntryType memberUpdate = AuditLogEntryType._create(24);
   static const AuditLogEntryType memberRoleUpdate = AuditLogEntryType._create(25);
+  static const AuditLogEntryType memberMove = AuditLogEntryType._create(26);
+  static const AuditLogEntryType memberDisconnect = AuditLogEntryType._create(27);
+  static const AuditLogEntryType botAdd = AuditLogEntryType._create(28);
   static const AuditLogEntryType roleCreate = AuditLogEntryType._create(30);
   static const AuditLogEntryType roleUpdate = AuditLogEntryType._create(31);
   static const AuditLogEntryType roleDelete = AuditLogEntryType._create(32);
@@ -106,6 +114,18 @@ class AuditLogEntryType extends IEnum<int> {
   static const AuditLogEntryType integrationCreate = AuditLogEntryType._create(80);
   static const AuditLogEntryType integrationUpdate = AuditLogEntryType._create(81);
   static const AuditLogEntryType integrationDelete = AuditLogEntryType._create(82);
+  static const AuditLogEntryType stageInstanceCreate = AuditLogEntryType._create(83);
+  static const AuditLogEntryType stageInstanceUpdate = AuditLogEntryType._create(84);
+  static const AuditLogEntryType stageInstanceDelete = AuditLogEntryType._create(85);
+  static const AuditLogEntryType stickerCreate = AuditLogEntryType._create(90);
+  static const AuditLogEntryType stickerUpdate = AuditLogEntryType._create(91);
+  static const AuditLogEntryType stickerDelete = AuditLogEntryType._create(92);
+  static const AuditLogEntryType guildScheduledEventCreate = AuditLogEntryType._create(100);
+  static const AuditLogEntryType guildScheduledEventUpdate = AuditLogEntryType._create(101);
+  static const AuditLogEntryType guildScheduledEventDelete = AuditLogEntryType._create(102);
+  static const AuditLogEntryType threadCreate = AuditLogEntryType._create(110);
+  static const AuditLogEntryType threadUpdate = AuditLogEntryType._create(111);
+  static const AuditLogEntryType threadDelete = AuditLogEntryType._create(112);
 
   const AuditLogEntryType._create(int value) : super(value);