+++ lib/src/core/message/components/message_component.dart
@@ -1,10 +1,10 @@
 import 'dart:convert';
 
-import 'package:nyxx/nyxx.dart';
 import 'package:nyxx/src/core/message/emoji.dart';
 import 'package:nyxx/src/core/message/guild_emoji.dart';
 import 'package:nyxx/src/core/message/unicode_emoji.dart';
 import 'package:nyxx/src/core/message/components/component_style.dart';
+import 'package:nyxx/src/core/snowflake.dart';
 import 'package:nyxx/src/typedefs.dart';
 import 'package:nyxx/src/utils/enum.dart';