+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## 4.0.0-dev.1
+__09.05.2022__
+
+- feature: Handle no internet on websocket (#321)
+- bug: Remove Error form IHttpResponseError (#324)
+  - Fixup field names on IHttpResponseError
+  - Fixup IHttpResponseSuccess name
+- feature: Move to API v10 (#325)
+
+## 4.0.0-dev.0
+__31.03.2022__
+
+- feature: Fix target id property and add guild audit logs options (#307)
+
+## 3.4.2
+__22.04.2022__
+
+- bug: Fix setting `channel` to `null` in MemberBuilder causing errors
+
 ## 3.4.1
 __10.04.2022__